Master upload
This application was designed to communicate with a server hosted by our
university.  The app represents a 4 (or more) way traffic light
controlled intersection. The input dictates the cars and pedestrians
coming into the intersection, indicating which exit they are coming from
and which exit they would like to take. The application sorts the cars
and pedestrians into their destination lanes.